Chore: move constants to own file
