+++ public/app/features/variables/constants.ts
@@ -1,0 +1,5 @@
+export const NEW_VARIABLE_ID = '00000000-0000-0000-0000-000000000000';
+export const ALL_VARIABLE_TEXT = 'All';
+export const ALL_VARIABLE_VALUE = '$__all';
+export const NONE_VARIABLE_TEXT = 'None';
+export const NONE_VARIABLE_VALUE = '';
