+++ public/app/features/variables/pickers/OptionsPicker/reducer.test.ts
@@ -16,7 +16,7 @@ import {
 } from './reducer';
 import { reducerTester } from '../../../../../test/core/redux/reducerTester';
 import { QueryVariableModel, VariableOption } from '../../types';
-import { ALL_VARIABLE_TEXT, ALL_VARIABLE_VALUE } from '../../state/types';
+import { ALL_VARIABLE_TEXT, ALL_VARIABLE_VALUE } from '../../constants';
 
 const getVariableTestContext = (extend: Partial<OptionsPickerState>) => {
   return {
