+++ public/app/features/variables/pickers/OptionsPicker/reducer.ts
@@ -1,10 +1,10 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { cloneDeep, isString, trim } from 'lodash';
 import { VariableOption, VariableWithOptions } from '../../types';
-import { ALL_VARIABLE_VALUE } from '../../state/types';
 import { isMulti, isQuery } from '../../guard';
 import { applyStateChanges } from '../../../../core/utils/applyStateChanges';
 import { containsSearchFilter } from '../../utils';
+import { ALL_VARIABLE_VALUE } from '../../constants';
 
 export interface ToggleOption {
   option?: VariableOption;
