+++ public/app/features/variables/query/reducer.ts
@@ -11,16 +11,8 @@ import {
   VariableSort,
 } from '../types';
 
-import {
-  ALL_VARIABLE_TEXT,
-  ALL_VARIABLE_VALUE,
-  getInstanceState,
-  initialVariablesState,
-  NONE_VARIABLE_TEXT,
-  NONE_VARIABLE_VALUE,
-  VariablePayload,
-  VariablesState,
-} from '../state/types';
+import { getInstanceState, initialVariablesState, VariablePayload, VariablesState } from '../state/types';
+import { ALL_VARIABLE_TEXT, ALL_VARIABLE_VALUE, NONE_VARIABLE_TEXT, NONE_VARIABLE_VALUE } from '../constants';
 
 interface VariableOptionsUpdate {
   templatedRegex: string;
